+++ src/main/kotlin/me/nobaboy/nobaaddons/features/qol/MouseLock.kt
@@ -3,13 +3,27 @@ package me.nobaboy.nobaaddons.features.qol
 import me.nobaboy.nobaaddons.api.SkyBlockAPI.inIsland
 import me.nobaboy.nobaaddons.config.NobaConfigManager
 import me.nobaboy.nobaaddons.core.SkyBlockIsland
+import me.nobaboy.nobaaddons.events.PacketEvents
 import me.nobaboy.nobaaddons.events.skyblock.SkyBlockEvents
+import me.nobaboy.nobaaddons.utils.LocationUtils
 import me.nobaboy.nobaaddons.utils.MCUtils
 import me.nobaboy.nobaaddons.utils.chat.ChatUtils
 import me.nobaboy.nobaaddons.utils.items.ItemUtils.getSkyBlockItem
+import me.nobaboy.nobaaddons.utils.toNobaVec
+import net.minecraft.network.packet.s2c.play.PlayerPositionLookS2CPacket
 
 object MouseLock {
+	val config = NobaConfigManager.config.qol.garden
+
 	private val FARMING_TOOLS: List<String> = buildList {
+		val gardeningTools = listOf("HOE", "AXE")
+		val gardeningToolTiers = listOf("BASIC", "ADVANCED")
+		gardeningToolTiers.forEach { tier ->
+			gardeningTools.forEach { tool ->
+				add("${tier}_GARDENING_$tool")
+			}
+		}
+
 		val theoreticalHoeCrops = listOf("WHEAT", "CARROT", "POTATO", "WARTS", "CANE")
 		val dicers = listOf("MELON_DICER", "PUMPKIN_DICER")
 		(1..3).forEach { tier ->
@@ -31,8 +45,6 @@ object MouseLock {
 	val reduced: Boolean get() {
 		if(!SkyBlockIsland.GARDEN.inIsland()) return false
 		if(MCUtils.player?.abilities?.flying == true) return false
-
-		val config = NobaConfigManager.config.qol.garden
 		if(!config.reduceMouseSensitivity) return false
 
 		val heldItem = MCUtils.player?.mainHandStack?.getSkyBlockItem() ?: return false
@@ -42,6 +54,19 @@ object MouseLock {
 
 	fun init() {
 		SkyBlockEvents.ISLAND_CHANGE.register { locked = false }
+		PacketEvents.EARLY_RECEIVE.register(this::onEarlyPacketReceive)
+	}
+
+	private fun onEarlyPacketReceive(event: PacketEvents.EarlyReceive) {
+		if(!config.autoUnlockMouseOnTeleport) return
+		if(!locked) return
+
+		val packet = event.packet
+		if(packet !is PlayerPositionLookS2CPacket) return
+
+		val playerLocation = LocationUtils.playerLocation().round(2)
+		val packetLocation = packet.change.position.toNobaVec().round(2)
+		if(packetLocation.distance(playerLocation) >= 5) locked = false
 	}
 
 	fun lockMouse() {